DEV: adjust default plotting lib to none
This commit makes an adjustment to the configuration file that
stores the default/fallback configuration settings that are used
when no settings are found in the User config file, user.cfg.

Specifically, the setting for default plotting library is changed
to `none`, from the previous value of `matplotlib`. This change
allows the `matplotlib` package to be truly treated as an extra
or optional dependency for GPy, and addresses GH Issue #873.

A further sequence of commits will explore updating the
documentation so that users who want to make use of the plotting
utilities of GPy can easily find the setting they need to change
to enable usage of a plotting library such as `matplotib` or
`plotly`.